create v0/bootstrap route
