+++ nexus-api/src/routes/v0/endpoints.rs
@@ -9,7 +9,6 @@ pub const INFO_ROUTE: &str = concatcp!(VERSION_ROUTE, "/info");
 // -- USER endpoints --
 const USER_PREFIX: &str = concatcp!(VERSION_ROUTE, "/user");
 pub const USER_ROUTE: &str = concatcp!(USER_PREFIX, "/{user_id}");
-pub const USER_BOOTSTRAP_ROUTE: &str = concatcp!(USER_ROUTE, "/im_alive");
 pub const RELATIONSHIP_ROUTE: &str = concatcp!(USER_ROUTE, "/relationship/{viewer_id}");
 pub const USER_COUNTS_ROUTE: &str = concatcp!(USER_ROUTE, "/counts");
 pub const USER_DETAILS_ROUTE: &str = concatcp!(USER_ROUTE, "/details");
@@ -49,17 +48,19 @@ const SEARCH_PREFIX: &str = concatcp!(VERSION_ROUTE, "/search");
 pub const SEARCH_USERS_ROUTE: &str = concatcp!(SEARCH_PREFIX, "/users");
 pub const SEARCH_TAGS_ROUTE: &str = concatcp!(SEARCH_PREFIX, "/tags/{label}");
 
-// TAG endpoints
-// Axum routes
+// -- TAG endpoints --
 const TAG_PREFIX: &str = concatcp!(VERSION_ROUTE, "/tags");
 pub const TAGS_HOT_ROUTE: &str = concatcp!(TAG_PREFIX, "/hot");
 pub const TAG_TAGGERS_ROUTE: &str = concatcp!(TAG_PREFIX, "/taggers/{label}");
 
-// FILE endpoints
-// Axum routes
+// -- FILE endpoints --
 const FILE_PREFIX: &str = concatcp!(VERSION_ROUTE, "/files");
 pub const FILE_LIST_ROUTE: &str = concatcp!(FILE_PREFIX, "/by_ids");
 pub const FILE_ROUTE: &str = concatcp!(FILE_PREFIX, "/file/{file_id}");
 
-// Notification route
+// -- NOTIFICATION endpoints -
 pub const NOTIFICATION_ROUTE: &str = concatcp!(USER_ROUTE, "/notifications");
+
+// -- BOOTSTRAP endpoints -
+const BOOTSTRAP_PREFIX: &str = concatcp!(VERSION_ROUTE, "/bootstrap");
+pub const IM_ALIVE_ROUTE: &str = concatcp!(BOOTSTRAP_PREFIX, "/{user_id}/im_alive");
