+++ nexus-api/src/routes/v0/user/mod.rs
@@ -15,7 +15,6 @@ mod view;
 pub fn routes() -> Router<AppState> {
     register_routes!(Router::new(),
         endpoints::USER_ROUTE => view::user_view_handler,
-        endpoints::USER_BOOTSTRAP_ROUTE => view::user_bootstrap_handler,
         endpoints::USER_DETAILS_ROUTE => details::user_details_handler,
         endpoints::RELATIONSHIP_ROUTE => relationship::user_relationship_handler,
         endpoints::USER_TAGS_ROUTE => tags::user_tags_handler,
