+++ nexus-api/src/routes/v0/user/view.rs
@@ -1,8 +1,7 @@
-use crate::routes::v0::endpoints::{USER_BOOTSTRAP_ROUTE, USER_ROUTE};
+use crate::routes::v0::endpoints::USER_ROUTE;
 use crate::{Error, Result};
 use axum::extract::{Path, Query};
 use axum::Json;
-use nexus_common::models::bootstrap::{Bootstrap, ViewType};
 use nexus_common::models::tag::TagDetails;
 use nexus_common::models::user::UserView;
 use serde::Deserialize;
@@ -47,40 +46,6 @@ pub async fn user_view_handler(
     }
 }
 
-#[utoipa::path(
-    get,
-    path = USER_BOOTSTRAP_ROUTE,
-    description = "Aggregated initial payload of all data required to bootstrap the client application",
-    tag = "User",
-    params(
-        ("user_id" = String, Path, description = "User Pubky ID")
-    ),
-    responses(
-        (status = 200, description = "Initial payload to bootstrap the client", body = Bootstrap),
-        (status = 404, description = "User payload not found"),
-        (status = 500, description = "Internal server error")
-    )
-)]
-pub async fn user_bootstrap_handler(
-    Path(user_id): Path<String>,
-    // TODO: Might need to add param like "ViewType". There are some data that it would be too much to delete in the first go
-    //Query(query): Query<Pub>,
-) -> Result<Json<Bootstrap>> {
-    info!("GET {USER_BOOTSTRAP_ROUTE} user_id:{}", user_id);
-
-    let view_type = ViewType::Full;
-
-    match Bootstrap::build(&user_id, view_type).await {
-        Ok(result) => Ok(Json(result)),
-        Err(source) => Err(Error::InternalServerError { source }),
-    }
-    // println!("Until new client, im_alive endpoint not available. NOTE: If you want to test, uncomment above code ;)");
-    // Ok(Json(Bootstrap::default()))
-}
-
 #[derive(OpenApi)]
-#[openapi(
-    paths(user_view_handler, user_bootstrap_handler),
-    components(schemas(UserView, TagDetails, Bootstrap))
-)]
+#[openapi(paths(user_view_handler), components(schemas(UserView, TagDetails)))]
 pub struct UserViewApiDoc;
